Old_Code/requirements.txt
numpy==1.23.5
pandas==1.4.4
yfinance==0.2.16
requests==2.28.2
cvxpy==1.2.3
statsmodels==0.13.5
pytz==2022.7
typing-extensions==4.5.0
python-dotenv==0.20.0
beautifulsoup4==4.11.1
aiohttp==3.8.1
scipy==1.10.1
streamlit==1.37.0